Fixed type casting
nx.relabel_graph was converting the string names to string integer ids. Needed to add an addition nx.relabel_graph to then convert those string integers to type ints.

Also added an assertion to check that the original graph is recoverable with nx.relabel_graph to check that the string to int id conversion is correct.