Updated configs and tasks that generate files
Configs have a warning flag to in case user wants a warning to execute that config e.g. full.ini has 31 tasks and do not want to accidentally run that.

The three tasks in this commit also no create task-specific folders for their outputs instead of cluttering the main output dir.